Split Animation into SequenceAnimation and ClipAnimation, retiring the nullable Clip discriminator
